bench/v4/tasks/1-glide-multilang/workspace/stats-python/requirements.txt
valkey-glide>=2.3.0
pytest>=8.0.0
pytest-asyncio>=0.24.0
